requirements.txt
google-api-python-client==2.170.0
pandas==2.2.3
python-dotenv==1.1.0
